+++ src/interfaces.ts
@@ -1,20 +1,6 @@
-// export interface DecodedToken {
-//   iss: number;
-//   exp: number;
-//   nbf: number;
-//   [key: string]: any;
-// }
 
 export interface RowWithTokens {
   accessToken?: string;
   refreshToken?: string;
   [key: string]: any;
 }
-
-
-// export interface ResponseLocalsWithTokens {
-//   id?: number;
-//   accessToken?: string;
-//   refreshToken?: string;
-//   [key: string]: any;
-// }